Move initializers to their own files
