+++ src/Rhisis.World/Systems/Quest/QuestSystem.cs
@@ -1,20 +1,14 @@
-﻿using Microsoft.EntityFrameworkCore;
-using Microsoft.EntityFrameworkCore.Metadata.Conventions;
-using Microsoft.Extensions.Logging;
+﻿using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
 using Rhisis.Core.Common;
 using Rhisis.Core.Data;
 using Rhisis.Core.DependencyInjection;
 using Rhisis.Core.Helpers;
 using Rhisis.Core.Resources;
-using Rhisis.Core.Structures.Configuration.Models;
 using Rhisis.Core.Structures.Configuration.World;
 using Rhisis.Core.Structures.Game;
 using Rhisis.Core.Structures.Game.Dialogs;
 using Rhisis.Core.Structures.Game.Quests;
-using Rhisis.Database;
-using Rhisis.Database.Entities;
-using Rhisis.World.Game.Components;
 using Rhisis.World.Game.Entities;
 using Rhisis.World.Game.Factories;
 using Rhisis.World.Game.Structures;
@@ -50,7 +44,6 @@ namespace Rhisis.World.Systems.Quest
         };
 
         private readonly ILogger<QuestSystem> _logger;
-        private readonly IRhisisDatabase _database;
         private readonly IGameResources _gameResources;
         private readonly WorldConfiguration _worldConfiguration;
         private readonly IPlayerDataSystem _playerDataSystem;
@@ -68,13 +61,12 @@ namespace Rhisis.World.Systems.Quest
         /// <inheritdoc />
         public int Order => 3;
 
-        public QuestSystem(ILogger<QuestSystem> logger, IRhisisDatabase database, IGameResources gameResources, IOptions<WorldConfiguration> worldServerConfiguration,
+        public QuestSystem(ILogger<QuestSystem> logger, IGameResources gameResources, IOptions<WorldConfiguration> worldServerConfiguration,
             IPlayerDataSystem playerDataSystem, IInventorySystem inventorySystem, IExperienceSystem experienceSystem, IJobSystem jobSystem,
             IStatisticsSystem statisticsSystem, ISkillSystem skillSystem, IDropSystem dropSystem,
             IQuestPacketFactory questPacketFactory, INpcDialogPacketFactory npcDialogPacketFactory, ITextPacketFactory textPacketFactory, IItemFactory itemFactory)
         {
             _logger = logger;
-            _database = database;
             _gameResources = gameResources;
             _worldConfiguration = worldServerConfiguration.Value;
             _playerDataSystem = playerDataSystem;
@@ -88,92 +80,6 @@ namespace Rhisis.World.Systems.Quest
             _npcDialogPacketFactory = npcDialogPacketFactory;
             _textPacketFactory = textPacketFactory;
             _itemFactory = itemFactory;
-        }
-
-        /// <inheritdoc />
-        public void Initialize(IPlayerEntity player)
-        {
-            IEnumerable<QuestInfo> playerQuests = _database.Quests.Where(x => x.CharacterId == player.PlayerData.Id)
-                .AsNoTracking()
-                .AsEnumerable()
-                .Select(x =>
-                {
-                    IQuestScript questScript = _gameResources.Quests.GetValueOrDefault(x.QuestId);
-
-                    if (questScript == null)
-                    {
-                        return null;
-                    }
-
-                    var quest = new QuestInfo(x.QuestId, x.CharacterId, questScript, x.Id)
-                    {
-                        IsChecked = x.IsChecked,
-                        IsFinished = x.Finished,
-                        StartTime = x.StartTime,
-                        IsPatrolDone = x.IsPatrolDone
-                    };
-
-                    if (questScript.EndConditions.Monsters != null && questScript.EndConditions.Monsters.Any())
-                    {
-                        quest.Monsters = new Dictionary<int, short>
-                        {
-                            { _gameResources.GetDefinedValue(questScript.EndConditions.Monsters.ElementAtOrDefault(0)?.Id), (short)x.MonsterKilled1 },
-                            { _gameResources.GetDefinedValue(questScript.EndConditions.Monsters.ElementAtOrDefault(1)?.Id), (short)x.MonsterKilled2 }
-                        };
-                    }
-
-                    return quest;
-                })
-                .Where(x => x != null);
-
-            player.QuestDiary = new QuestDiaryComponent(playerQuests);
-        }
-
-        /// <inheritdoc />
-        public void Save(IPlayerEntity player)
-        {
-            var questsSet = from x in _database.Quests.Where(x => x.CharacterId == player.PlayerData.Id).ToList()
-                            join q in player.QuestDiary on
-                             new { x.QuestId, x.CharacterId }
-                             equals
-                             new { q.QuestId, q.CharacterId }
-                            select new { DbQuest = x, PlayerQuest = q };
-
-            foreach (var questSet in questsSet)
-            {
-                questSet.DbQuest.IsChecked = questSet.PlayerQuest.IsFinished ? false : questSet.PlayerQuest.IsChecked;
-                questSet.DbQuest.IsDeleted = questSet.PlayerQuest.IsDeleted;
-                questSet.DbQuest.IsPatrolDone = questSet.PlayerQuest.IsPatrolDone;
-                questSet.DbQuest.Finished = questSet.PlayerQuest.IsFinished;
-
-                if (questSet.PlayerQuest.Monsters != null)
-                {
-                    questSet.DbQuest.MonsterKilled1 = questSet.PlayerQuest.Monsters.ElementAtOrDefault(0).Value;
-                    questSet.DbQuest.MonsterKilled2 = questSet.PlayerQuest.Monsters.ElementAtOrDefault(1).Value;
-                }
-
-                _database.Quests.Update(questSet.DbQuest);
-            }
-
-            foreach (QuestInfo quest in player.QuestDiary)
-            {
-                if (!quest.DatabaseQuestId.HasValue)
-                {
-                    _database.Quests.Add(new DbQuest
-                    {
-                        CharacterId = player.PlayerData.Id,
-                        QuestId = quest.QuestId,
-                        StartTime = quest.StartTime,
-                        MonsterKilled1 = quest.Monsters?.ElementAtOrDefault(0).Value ?? default,
-                        MonsterKilled2 = quest.Monsters?.ElementAtOrDefault(1).Value ?? default,
-                        IsPatrolDone = quest.IsPatrolDone,
-                        IsChecked = quest.IsFinished ? false : quest.IsChecked,
-                        Finished = quest.IsFinished
-                    });
-                }
-            }
-
-            _database.SaveChanges();
         }
 
         /// <inheritdoc />
